+++ TriviaServer/TriviaServer/DAO/Repositories/PlayerRepository.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -17,29 +18,51 @@ namespace TriviaServer.DAO.Repositories
             _context = context;
         }
 
-        public void Create(Player p)
+        public void Create(Player player)
         {
-            _context.Players.Add(p);
-            _context.SaveChanges();
+            var players = _context.Players.Where(a => a.GameroomId == player.GameroomId).ToList();
+            if (Validation.CreatePlayerValidation(players,player))
+            {
+                _context.Players.Add(player);
+                _context.SaveChanges();
+            }
+            else
+            {
+                throw new Exception("Nume deja existent!");
+            }
         }
 
-        public void Edit(Player p)
+        public void Edit(Player player)
         {
-            _context.Players.Update(p);
+            _context.Players.Update(player);
             _context.SaveChanges();
         }
 
         public void Delete(int id)
         {
             var player = _context.Players.SingleOrDefault(x => x.PlayerId == id);
-            _context.Players.Remove(player);
-            _context.SaveChanges();
+            if(player != null)
+            {
+                _context.Players.Remove(player);
+                _context.SaveChanges();
+            }
+            else
+            {
+                throw new Exception("Player not found!");
+            }
         }
 
         public IEnumerable<Player> GetPlayers()
         {
             var players = _context.Players.ToList();
-            return players;
+            if(players != null)
+            {
+                return players;
+            }
+            else
+            {
+                throw new Exception("No player was found!");
+            }
         }
 
         public Player GetByID(int? id)
@@ -48,18 +71,31 @@ namespace TriviaServer.DAO.Repositories
             {
                 return null;
             }
-
             var player = _context.Players.SingleOrDefault(p => p.PlayerId == id);
-
-            return player;
+            if (player != null)
+            {
+                return player;
+            }
+            else
+            {
+                throw new Exception("Player not found!");
+            }
         }
 
         public void UpdatePlayerScore(int gameRoomId, String playerName, int score)
         {
-            var player = _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
+            var players = _context.Players.Where(a => a.GameroomId == gameRoomId).ToList();
+            if (Validation.UpdateScoreValidation(players, playerName))
+            {
+                var player = _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
                 .Where(a => a.PlayerName == playerName).First();
-            player.PlayerScore += score;
-            Edit(player);
+                player.PlayerScore += score;
+                Edit(player);
+            }
+            else
+            {
+                throw new Exception("Player not found!");
+            }
         }
     }
 }
